Time: 3 ms (94.03%), Space: 61.8 MB (52.06%) - LeetHub

+++ 2432-number-of-zero-filled-subarrays/2432-number-of-zero-filled-subarrays.java
@@ -1,7 +1,7 @@
 class Solution {
     public long zeroFilledSubarray(int[] nums) {
         long result = 0;
-        long indexOfzeroes = 0;
+        int indexOfzeroes = 0;
         for(int i=0; i<nums.length; i++) {
             if(nums[i] == 0) {
                 indexOfzeroes++;
